+++ packages/modals/src/views/Close.js
@@ -37,6 +37,8 @@ export default class Close extends Component {
     hovered: PropTypes.bool
   };
 
+  static hasType = () => Close;
+
   state = {
     isFocused: false
   };