+++ packages/select/src/views/items/Item.js
@@ -29,5 +29,7 @@ Item.propTypes = {
   checked: PropTypes.bool
 };
 
+Item.hasType = () => Item;
+
 /** @component */
 export default Item;