+++ packages/select/src/views/items/NextItem.js
@@ -28,5 +28,7 @@ NextItem.propTypes = {
   disabled: PropTypes.bool
 };
 
+NextItem.hasType = () => NextItem;
+
 /** @component */
 export default NextItem;