+++ packages/select/src/views/items/PreviousItem.js
@@ -28,5 +28,7 @@ PreviousItem.propTypes = {
   disabled: PropTypes.bool
 };
 
+PreviousItem.hasType = () => PreviousItem;
+
 /** @component */
 export default PreviousItem;