+++ packages/textfields/src/views/Hint.js
@@ -34,5 +34,7 @@ Hint.propTypes = {
   small: PropTypes.bool
 };
 
+Hint.hasType = () => Hint;
+
 /** @component */
 export default Hint;