+++ packages/textfields/src/views/Label.js
@@ -36,5 +36,7 @@ Label.propTypes = {
   small: PropTypes.bool
 };
 
+Label.hasType = () => Label;
+
 /** @component */
 export default Label;