+++ packages/toggles/src/views/Label.js
@@ -48,5 +48,7 @@ Label.propTypes = {
   disabled: PropTypes.bool
 };
 
+Label.hasType = () => Label;
+
 /** @component */
 export default Label;