+++ packages/chrome/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.0.3"></a>
+## [4.0.3](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-chrome@4.0.2...@zendeskgarden/react-chrome@4.0.3) (2018-06-06)
+
+
+
+
+**Note:** Version bump only for package @zendeskgarden/react-chrome
+
 <a name="4.0.2"></a>
 ## [4.0.2](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-chrome@4.0.1...@zendeskgarden/react-chrome@4.0.2) (2018-06-05)