+++ packages/tags/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tags@1.0.2...@zendeskgarden/react-tags@1.0.3) (2018-06-06)
+
+
+
+
+**Note:** Version bump only for package @zendeskgarden/react-tags
+
 <a name="1.0.2"></a>
 ## [1.0.2](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tags@1.0.1...@zendeskgarden/react-tags@1.0.2) (2018-06-05)